Prod/requirements.txt
fastapi[standard]
uvicorn
pydantic-settings
pydantic[email]
psycopg2-binary
PyJWT
httpx
pytest
pytest-asyncio
